Extract default GUI file actions
